Implement URL hashing API
Implemented a hashing API which relies on the Hashids library to generate URL hashes. It takes a URL as input and returns a short hash associated with the input URL. These hashes are used to identify the original URL, and are part of the shortened URL, e.g.: 'https://www.shortener.com/<hash>'.