Implement URL shortening API
Implemented a URL shortening API that takes a target_url as an argument and secures it before passing it to the hashing API to create a hash. Once a hash is generated, it is added to the database along its associated target_url. Finally, it is returned as a shortened URL, e.g.: 'https://www.linkshrink.app/<hash>'.